tests/wpt/tests/html/browsers/history/the-history-interface/history_pushstate_too_many_calls.optional.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>history pushState too many calls</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      assert_throws_dom("SecurityError", function () {
        for (let i = 0; i < 500; i++) {
          window.history.pushState(null, null, i);
        }
      });
    });
    </script>
  </body>
</html>
